Create barcs.py
Version 1.0.  (10/5/21)

I started this project for my wife who is looking to adopt either a basset hound or a corgi from the local animal shelter (BARCs, Baltimore Area Rescue Center). Future update will add the Name and Breed of the dog to the text, which is why the pet names are added as a dictionary.